+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/client/ByProjectKeyExtensionsKeyByKeyDelete.java
@@ -91,10 +91,6 @@ public class ByProjectKeyExtensionsKeyByKeyDelete extends
         return this.getQueryParam("version");
     }
 
-    public List<String> getExpand() {
-        return this.getQueryParam("expand");
-    }
-
     public void setProjectKey(final String projectKey) {
         this.projectKey = projectKey;
     }
@@ -180,85 +176,6 @@ public class ByProjectKeyExtensionsKeyByKeyDelete extends
     public <TValue> ByProjectKeyExtensionsKeyByKeyDelete addVersion(final Collection<TValue> version) {
         return copy().addQueryParams(
             version.stream().map(s -> new ParamEntry<>("version", s.toString())).collect(Collectors.toList()));
-    }
-
-    /**
-     * set expand with the specified value
-     * @param expand value to be set
-     * @param <TValue> value type
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public <TValue> ByProjectKeyExtensionsKeyByKeyDelete withExpand(final TValue expand) {
-        return copy().withQueryParam("expand", expand);
-    }
-
-    /**
-     * add additional expand query parameter
-     * @param expand value to be added
-     * @param <TValue> value type
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public <TValue> ByProjectKeyExtensionsKeyByKeyDelete addExpand(final TValue expand) {
-        return copy().addQueryParam("expand", expand);
-    }
-
-    /**
-     * set expand with the specified value
-     * @param supplier supplier for the value to be set
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public ByProjectKeyExtensionsKeyByKeyDelete withExpand(final Supplier<String> supplier) {
-        return copy().withQueryParam("expand", supplier.get());
-    }
-
-    /**
-     * add additional expand query parameter
-     * @param supplier supplier for the value to be added
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public ByProjectKeyExtensionsKeyByKeyDelete addExpand(final Supplier<String> supplier) {
-        return copy().addQueryParam("expand", supplier.get());
-    }
-
-    /**
-     * set expand with the specified value
-     * @param op builder for the value to be set
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public ByProjectKeyExtensionsKeyByKeyDelete withExpand(final Function<StringBuilder, StringBuilder> op) {
-        return copy().withQueryParam("expand", op.apply(new StringBuilder()));
-    }
-
-    /**
-     * add additional expand query parameter
-     * @param op builder for the value to be added
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public ByProjectKeyExtensionsKeyByKeyDelete addExpand(final Function<StringBuilder, StringBuilder> op) {
-        return copy().addQueryParam("expand", op.apply(new StringBuilder()));
-    }
-
-    /**
-     * set expand with the specified values
-     * @param expand values to be set
-     * @param <TValue> value type
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public <TValue> ByProjectKeyExtensionsKeyByKeyDelete withExpand(final Collection<TValue> expand) {
-        return copy().withoutQueryParam("expand")
-                .addQueryParams(
-                    expand.stream().map(s -> new ParamEntry<>("expand", s.toString())).collect(Collectors.toList()));
-    }
-
-    /**
-     * add additional expand query parameters
-     * @param expand values to be added
-     * @param <TValue> value type
-     * @return ByProjectKeyExtensionsKeyByKeyDelete
-     */
-    public <TValue> ByProjectKeyExtensionsKeyByKeyDelete addExpand(final Collection<TValue> expand) {
-        return copy().addQueryParams(
-            expand.stream().map(s -> new ParamEntry<>("expand", s.toString())).collect(Collectors.toList()));
     }
 
     @Override